misc fixes for erroneous diff found in backwards compatibility test (#2960)
<!-- CURSOR_SUMMARY -->
> [!NOTE]
> Adds semantic column/type equivalence (enums/JSON/nested), normalizes
TTLs by stripping DELETE, introduces `FixedString` to proto, improves
PY/TS recursive type collection, and standardizes default int/float to
Int64/Float64 with corresponding tests and regex escaping fixes.
> 
> - **Diffing/Equivalence**:
> - Implement semantic `column_types_are_equivalent` (enums, JSON
`typed_paths` order-independence, `Nested`), and use it in
`columns_are_equivalent`.
> - Add `nested_are_equivalent` and `json_options_are_equivalent`;
extensive tests for nested/JSON scenarios.
> - **TTL Handling**:
> - `normalize_ttl_expression` now strips trailing `DELETE`; update
tests accordingly.
> - **Proto/Types**:
> - Add `fixed_string` to `ColumnType` proto and map
`ColumnType::FixedString { length }` to/from proto.
> - **Codegen (Python/TypeScript)**:
> - Recursively collect types within `Nested`, `NamedTuple`,
`Json.typed_paths`, `Array`, `Nullable`, and `Map` to generate inner
models/interfaces.
> - **ClickHouse JSON Options**:
> - Correct SQL string escaping/unescaping for `SKIP REGEXP` in
render/parse paths.
> - **Language libs**:
> - Python: default `int`→`Int64`; avoid duplicate aggregation
annotations; update query param parsing to accept `Int64`.
>   - TypeScript: default number without tags→`Float64`.
> - **Tests/Templates**:
> - Update/expand tests for TTL normalization, JSON regex, semantic
equivalence, and default int/float mappings; adjust template JSON regex
literals.
> 
> <sup>Written by [Cursor
Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit
521b58f1ad006229418ae29ae932a4268157cbfe. This will update automatically
on new commits. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ packages/py-moose-lib/tests/test_int_types.py
@@ -183,7 +183,7 @@ def test_default_int_still_works():
     by_name = {col.name: col for col in columns}
 
     # Plain int should still map to "Int" (default behavior)
-    assert by_name["plain_int"].data_type == "Int"
+    assert by_name["plain_int"].data_type == "Int64"
     # Typed int should map to specific type
     assert by_name["typed_int"].data_type == "UInt32"
